# api/openapi.bundled.yaml
openapi: 3.2.0
...
info:
  title: Arches Platform API
  description: The Arches Platform API for managing projects and code generation
  version: v0.0.0
  summary: API for Arches platform
  termsOfService: https://archesai.com/terms
  license:
    name: AGPL-3.0
    url: https://opensource.org/licenses/AGPL-3.0
  contact:
    name: Arches AI Support
    email: support@archesai.com
    url: https://archesai.com/support
...
servers:
  - url: https://api.archesai.dev
    description: Remote development server
  - url: https://api.archesai.com
paths:
...
  /auth/logout-all:
    post:
      x-internal: auth
      operationId: LogoutAll
      x-codegen-custom-handler: true
      summary: Logout all sessions
      tags:
        - Auth
      description: Logout from all sessions across all devices
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        '200':
          $ref: '#/components/responses/LogoutResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /auth/link:
    post:
      x-internal: auth
      x-codegen-custom-handler: true
      operationId: LinkAccount
      summary: Link authentication provider
      tags:
        - Auth
      description: Link an additional authentication provider to the current user account
      security:
        - bearerAuth: []
        - sessionCookie: []
      requestBody:
        description: Account linking request
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - provider
              properties:
                provider:
                  description: The authentication provider to link
                  type: string
                  enum:
                    - google
                    - github
                    - microsoft
                    - apple
                  example: github
                redirectUrl:
                  description: URL to redirect to after successful linking
                  type: string
                  format: uri
                  maxLength: 2048
                  example: https://app.archesai.com/settings/security
      responses:
        '200':
          description: Provider linking initiated
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - authorizationUrl
                properties:
                  authorizationUrl:
                    description: URL to redirect the user to for provider authorization
                    type: string
                    format: uri
                    minLength: 1
                    maxLength: 2048
                    example: https://github.com/login/oauth/authorize?client_id=...
              example:
                authorizationUrl: https://github.com/login/oauth/authorize?client_id=abc123
          headers:
            X-RateLimit-Limit:
              description: The maximum number of requests allowed per time window
              schema:
                type: integer
                example: 1000
                format: int32
                minimum: 1
                maximum: 1000000
              example: 1000
            X-RateLimit-Remaining:
              description: The number of requests remaining in the current time window
              schema:
                type: integer
                example: 999
                format: int32
                minimum: 0
                maximum: 1000000
              example: 999
            X-RateLimit-Reset:
              description: The time at which the current rate limit window resets (Unix timestamp)
              schema:
                type: integer
                example: 1640995200
                format: int64
                minimum: 0
                maximum: 9223372036854776000
              example: 1640995200
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /organizations/{organizationID}/members/{id}:
    get:
      x-internal: auth
      operationId: GetMember
      summary: Get a member
      tags:
        - Member
      description: Get a member
      parameters:
        - $ref: '#/components/parameters/OrganizationID'
        - $ref: '#/components/parameters/ResourceID'
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/MemberResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      x-internal: auth
      operationId: UpdateMember
      summary: Update a member
      tags:
        - Member
      description: Update a member
      requestBody:
        description: Member update data
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  description: The role of the member
                  type: string
                  enum:
                    - admin
                    - owner
                    - member
                  example: admin
      parameters:
        - $ref: '#/components/parameters/OrganizationID'
        - $ref: '#/components/parameters/ResourceID'
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/MemberResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    delete:
      x-internal: auth
      operationId: DeleteMember
      summary: Delete a member
      tags:
        - Member
      description: Delete a member
      parameters:
        - $ref: '#/components/parameters/OrganizationID'
        - $ref: '#/components/parameters/ResourceID'
      security:
        - bearerAuth: []
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    Member:
      title: Member
      description: Schema for Member entity
      x-codegen-schema-type: entity
      x-codegen:
        repository:
          indices:
            - organizationID
            - userID
          excludeFromUpdate:
            - OrganizationID
            - UserID
          relations:
            - field: organizationID
              references: organization
              onDelete: CASCADE
              onUpdate: NO_ACTION
            - field: userID
              references: user
              onDelete: CASCADE
              onUpdate: NO_ACTION
          additionalMethods:
            - name: ListMembersByOrganization
              params:
                - name: organizationID
                  type: string
                  format: uuid
                  maxLength: 36
              returns: multiple
            - name: ListMembersByUser
              params:
                - name: userID
                  type: string
                  format: uuid
                  maxLength: 36
              returns: multiple
            - name: GetMemberByUserAndOrganization
              params:
                - name: userID
                  type: string
                  format: uuid
                  maxLength: 36
                - name: organizationID
                  type: string
                  format: uuid
                  maxLength: 36
              returns: single
      x-internal: auth
      allOf:
        - $ref: '#/components/schemas/Base'
        - type: object
          properties:
            organizationID:
              description: The organization this member belongs to
              type: string
              format: uuid
              example: 550e8400-e29b-41d4-a716-446655440000
              minLength: 36
              maxLength: 36
            role:
              description: The role of the member
              type: string
              enum:
                - admin
                - owner
                - basic
              example: admin
              default: basic
            userID:
              description: The user who is a member of the organization
              type: string
              format: uuid
              example: 550e8400-e29b-41d4-a716-446655440000
              minLength: 36
              maxLength: 36
          required:
            - organizationID
            - role
            - userID
      unevaluatedProperties: false
...
    Base:
      title: Base
      description: Base schema for all entities with common fields
      x-internal: server
      type: object
      properties:
        id:
          description: Unique identifier for the resource
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
          minLength: 36
          maxLength: 36
        createdAt:
          description: The date and time when the resource was created
          type: string
          format: date-time
          example: '2024-01-15T09:30:00Z'
          minLength: 1
          maxLength: 255
        updatedAt:
          description: The date and time when the resource was last updated
          type: string
          format: date-time
          example: '2024-01-15T10:45:00Z'
          minLength: 1
          maxLength: 255
      required:
        - id
        - createdAt
        - updatedAt
...
    UUID:
      description: UUID identifier
      x-internal: server
      type: string
      format: uuid
      example: 550e8400-e29b-41d4-a716-446655440000
      minLength: 36
      maxLength: 36
...
    Problem:
      title: Problem
      description: RFC 7807 (Problem Details) compliant error response
      x-internal: server
      type: object
      additionalProperties: false
      required:
        - status
        - title
      properties:
        status:
          type: integer
          format: int32
          description: HTTP status code
          minimum: 100
          maximum: 599
          example: 400
        type:
          type: string
          format: uri
          description: URI identifying the problem type
          default: about:blank
          example: https://api.example.com/errors/validation-failed
          minLength: 1
          maxLength: 2048
        title:
          type: string
          description: Short, human-readable summary
          example: Validation Failed
          minLength: 1
          maxLength: 255
          pattern: ^[\w\s\-.,!?()]+$
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence
          example: The request body contains invalid fields
          minLength: 1
          maxLength: 255
          pattern: ^[\w\s\-.,!?()]+$
        instance:
          type: string
          format: uri
          description: URI identifying the specific occurrence
          example: https://api.example.com/auth/users/123
          maxLength: 2048
...
  parameters:
...
    ResourceID:
      schema:
        $ref: '#/components/schemas/UUID'
      in: path
      name: id
      required: true
      description: The unique identifier of the resource.
...
    OrganizationID:
      schema:
        $ref: '#/components/schemas/UUID'
      in: path
      name: organizationID
      required: true
      description: The unique identifier of the organization.
...
  responses:
    BadRequest:
      description: 400 Bad Request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
    Conflict:
      description: 409 Conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
...
    InternalServerError:
      description: Internal server error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
    NoContent:
      description: 204 No Content
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
    NotFound:
      description: 404 Not Found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
    TooManyRequests:
      description: Too many requests - rate limit exceeded
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
    Unauthorized:
      description: 401 Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
    UnprocessableEntity:
      description: 422 Unprocessable Entity
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
...
    LogoutResponse:
      description: Logout successful
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required:
              - message
            properties:
              message:
                type: string
                minLength: 1
                maxLength: 255
                example: Successfully logged out
                pattern: ^[\w\s\-.,!?()]+$
          example:
            message: Successfully logged out
      headers:
        Set-Cookie:
          $ref: '#/components/headers/SetCookie'
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
...
    MemberResponse:
      description: Member retrieved successfully
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                $ref: '#/components/schemas/Member'
            required:
              - data
            additionalProperties: false
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
...
  headers:
    RateLimitLimit:
      description: The maximum number of requests allowed per time window
      schema:
        type: integer
        example: 1000
        format: int32
        minimum: 1
        maximum: 1000000
      example: 1000
    RateLimitRemaining:
      description: The number of requests remaining in the current time window
      schema:
        type: integer
        example: 999
        format: int32
        minimum: 0
        maximum: 1000000
      example: 999
    RateLimitReset:
      description: The time at which the current rate limit window resets (Unix timestamp)
      schema:
        type: integer
        example: 1640995200
        format: int64
        minimum: 0
        maximum: 9223372036854776000
      example: 1640995200
    RetryAfter:
      description: Number of seconds to wait before making a new request
      schema:
        type: integer
        example: 60
        format: int32
        minimum: 1
        maximum: 3600
      example: 60
    SetCookie:
      description: Session cookies for authentication
      schema:
        type: string
        minLength: 1
        maxLength: 4096
        pattern: ^[\w\-=;,. ]+$
  securitySchemes:
    bearerAuth:
      description: API Token for authenticated requests
      scheme: bearer
      type: http
    sessionCookie:
      description: Session cookie for authenticated requests
      type: apiKey
      name: __Secure-better-auth.session_token
      in: cookie
security:
  - bearerAuth: []
  - sessionCookie: []