# pkg/auth/gen/controllers/link_account.gen.go
// Code generated by archesai. DO NOT EDIT.

package controllers

import (
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/google/uuid"

	"github.com/archesai/archesai/pkg/auth/gen/application"
	"github.com/archesai/archesai/pkg/server"
)

// ============================================================================
// LinkAccount - POST /auth/link
// ============================================================================

// LinkAccountHandler is the HTTP handler for LinkAccount.
type LinkAccountHandler struct {
	linkAccount application.LinkAccount
}

// NewLinkAccountHandler creates a new HTTP handler.
func NewLinkAccountHandler(linkAccount application.LinkAccount) *LinkAccountHandler {
	return &LinkAccountHandler{linkAccount: linkAccount}
}

// RegisterLinkAccountRoute registers the HTTP route for LinkAccount.
func RegisterLinkAccountRoute(mux *http.ServeMux, handler *LinkAccountHandler) {
	mux.HandleFunc("POST /auth/link", handler.ServeHTTP)
}

// Request types

// LinkAccountRequestBody defines the request body for LinkAccount
type LinkAccountRequestBody struct {
	Provider    string  `json:"provider"`
	RedirectURL *string `json:"redirectUrl,omitempty"`
}

// Response types

type LinkAccountResponse interface {
	VisitLinkAccountResponse(w http.ResponseWriter) error
}

type LinkAccount200Response struct {
	AuthorizationURL string `json:"authorizationURL"`
}

func (response LinkAccount200Response) VisitLinkAccountResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	return json.NewEncoder(w).Encode(response)
}

type LinkAccount400Response struct {
	server.ProblemDetails
}

func (response LinkAccount400Response) VisitLinkAccountResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(400)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type LinkAccount401Response struct {
	server.ProblemDetails
}

func (response LinkAccount401Response) VisitLinkAccountResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(401)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type LinkAccount409Response struct {
	server.ProblemDetails
}

func (response LinkAccount409Response) VisitLinkAccountResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(409)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type LinkAccount422Response struct {
	server.ProblemDetails
}

func (response LinkAccount422Response) VisitLinkAccountResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(422)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type LinkAccount429Response struct {
	server.ProblemDetails
}

func (response LinkAccount429Response) VisitLinkAccountResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(429)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type LinkAccount500Response struct {
	server.ProblemDetails
}

func (response LinkAccount500Response) VisitLinkAccountResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(500)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

// ServeHTTP handles the POST /auth/link endpoint.
func (h *LinkAccountHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	// Extract session ID from context for authenticated operations
	sessionID, ok := ctx.Value(server.SessionIDContextKey).(uuid.UUID)
	if !ok {
		errorResp := LinkAccount401Response{
			ProblemDetails: server.NewUnauthorizedResponse("session required", r.URL.Path),
		}
		if err := errorResp.VisitLinkAccountResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}

	// Build input from request
	input := &application.LinkAccountInput{}
	input.SessionID = sessionID

	// Request body
	body := &LinkAccountRequestBody{}
	if err := json.NewDecoder(r.Body).Decode(body); err != nil {
		errorResp := LinkAccount400Response{
			ProblemDetails: server.NewBadRequestResponse(err.Error(), r.URL.Path),
		}
		if err := errorResp.VisitLinkAccountResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}
	input.Provider = body.Provider
	input.RedirectURL = body.RedirectURL

	// Execute
	result, err := h.linkAccount.Execute(ctx, input)
	if err != nil {
		errorResp := LinkAccount500Response{
			ProblemDetails: server.NewInternalServerErrorResponse(err.Error(), r.URL.Path),
		}
		if err := errorResp.VisitLinkAccountResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}

	// Map output to response
	response := LinkAccount200Response{}
	response.AuthorizationURL = result.AuthorizationURL

	if err := response.VisitLinkAccountResponse(w); err != nil {
		fmt.Fprintf(w, "error writing response: %v", err)
	}
}