# pkg/auth/gen/controllers/logout_all.gen.go
// Code generated by archesai. DO NOT EDIT.

package controllers

import (
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/google/uuid"

	"github.com/archesai/archesai/pkg/auth/gen/application"
	"github.com/archesai/archesai/pkg/server"
)

// ============================================================================
// LogoutAll - POST /auth/logout-all
// ============================================================================

// LogoutAllHandler is the HTTP handler for LogoutAll.
type LogoutAllHandler struct {
	logoutAll application.LogoutAll
}

// NewLogoutAllHandler creates a new HTTP handler.
func NewLogoutAllHandler(logoutAll application.LogoutAll) *LogoutAllHandler {
	return &LogoutAllHandler{logoutAll: logoutAll}
}

// RegisterLogoutAllRoute registers the HTTP route for LogoutAll.
func RegisterLogoutAllRoute(mux *http.ServeMux, handler *LogoutAllHandler) {
	mux.HandleFunc("POST /auth/logout-all", handler.ServeHTTP)
}

// Request types

// Response types

type LogoutAllResponse interface {
	VisitLogoutAllResponse(w http.ResponseWriter) error
}

type LogoutAll200Response struct {
	Message string `json:"message"`
}

func (response LogoutAll200Response) VisitLogoutAllResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	return json.NewEncoder(w).Encode(response)
}

type LogoutAll400Response struct {
	server.ProblemDetails
}

func (response LogoutAll400Response) VisitLogoutAllResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(400)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type LogoutAll401Response struct {
	server.ProblemDetails
}

func (response LogoutAll401Response) VisitLogoutAllResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(401)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type LogoutAll429Response struct {
	server.ProblemDetails
}

func (response LogoutAll429Response) VisitLogoutAllResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(429)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type LogoutAll500Response struct {
	server.ProblemDetails
}

func (response LogoutAll500Response) VisitLogoutAllResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(500)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

// ServeHTTP handles the POST /auth/logout-all endpoint.
func (h *LogoutAllHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	// Extract session ID from context for authenticated operations
	sessionID, ok := ctx.Value(server.SessionIDContextKey).(uuid.UUID)
	if !ok {
		errorResp := LogoutAll401Response{
			ProblemDetails: server.NewUnauthorizedResponse("session required", r.URL.Path),
		}
		if err := errorResp.VisitLogoutAllResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}

	// Build input from request
	input := &application.LogoutAllInput{}
	input.SessionID = sessionID

	// Execute
	result, err := h.logoutAll.Execute(ctx, input)
	if err != nil {
		errorResp := LogoutAll500Response{
			ProblemDetails: server.NewInternalServerErrorResponse(err.Error(), r.URL.Path),
		}
		if err := errorResp.VisitLogoutAllResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}

	// Map output to response
	response := LogoutAll200Response{}
	response.Message = result.Message

	if err := response.VisitLogoutAllResponse(w); err != nil {
		fmt.Fprintf(w, "error writing response: %v", err)
	}
}