# pkg/auth/gen/controllers/update_member.gen.go
// Code generated by archesai. DO NOT EDIT.

package controllers

import (
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/google/uuid"
	"github.com/oapi-codegen/runtime"

	"github.com/archesai/archesai/pkg/auth/gen/application"
	"github.com/archesai/archesai/pkg/auth/gen/models"
	"github.com/archesai/archesai/pkg/server"
)

// ============================================================================
// UpdateMember - PATCH /organizations/{organizationID}/members/{id}
// ============================================================================

// UpdateMemberHandler is the HTTP handler for UpdateMember.
type UpdateMemberHandler struct {
	updateMember application.UpdateMember
}

// NewUpdateMemberHandler creates a new HTTP handler.
func NewUpdateMemberHandler(updateMember application.UpdateMember) *UpdateMemberHandler {
	return &UpdateMemberHandler{updateMember: updateMember}
}

// RegisterUpdateMemberRoute registers the HTTP route for UpdateMember.
func RegisterUpdateMemberRoute(mux *http.ServeMux, handler *UpdateMemberHandler) {
	mux.HandleFunc("PATCH /organizations/{organizationID}/members/{id}", handler.ServeHTTP)
}

// Request types

// UpdateMemberRequestBody defines the request body for UpdateMember
type UpdateMemberRequestBody struct {
	Role *string `json:"role,omitempty"`
}

// Response types

type UpdateMemberResponse interface {
	VisitUpdateMemberResponse(w http.ResponseWriter) error
}

type UpdateMember200Response struct {
	Data models.Member `json:"data"`
}

func (response UpdateMember200Response) VisitUpdateMemberResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	return json.NewEncoder(w).Encode(response)
}

type UpdateMember400Response struct {
	server.ProblemDetails
}

func (response UpdateMember400Response) VisitUpdateMemberResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(400)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type UpdateMember401Response struct {
	server.ProblemDetails
}

func (response UpdateMember401Response) VisitUpdateMemberResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(401)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type UpdateMember404Response struct {
	server.ProblemDetails
}

func (response UpdateMember404Response) VisitUpdateMemberResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(404)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type UpdateMember422Response struct {
	server.ProblemDetails
}

func (response UpdateMember422Response) VisitUpdateMemberResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(422)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type UpdateMember429Response struct {
	server.ProblemDetails
}

func (response UpdateMember429Response) VisitUpdateMemberResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(429)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type UpdateMember500Response struct {
	server.ProblemDetails
}

func (response UpdateMember500Response) VisitUpdateMemberResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(500)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

// ServeHTTP handles the PATCH /organizations/{organizationID}/members/{id} endpoint.
func (h *UpdateMemberHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	// Extract session ID from context for authenticated operations
	sessionID, ok := ctx.Value(server.SessionIDContextKey).(uuid.UUID)
	if !ok {
		errorResp := UpdateMember401Response{
			ProblemDetails: server.NewUnauthorizedResponse("session required", r.URL.Path),
		}
		if err := errorResp.VisitUpdateMemberResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}

	// Build input from request
	input := &application.UpdateMemberInput{}
	input.SessionID = sessionID

	// Path parameter "organizationID"
	var organizationID uuid.UUID
	if err := runtime.BindStyledParameterWithOptions("simple", "organizationID", r.PathValue("organizationID"), &organizationID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true}); err != nil {
		errorResp := UpdateMember400Response{
			ProblemDetails: server.NewBadRequestResponse(fmt.Sprintf("Invalid format for parameter organizationID: %s", err), r.URL.Path),
		}
		if err := errorResp.VisitUpdateMemberResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}
	input.OrganizationID = organizationID

	// Path parameter "id"
	var id uuid.UUID
	if err := runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true}); err != nil {
		errorResp := UpdateMember400Response{
			ProblemDetails: server.NewBadRequestResponse(fmt.Sprintf("Invalid format for parameter id: %s", err), r.URL.Path),
		}
		if err := errorResp.VisitUpdateMemberResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}
	input.ID = id

	// Request body
	body := &UpdateMemberRequestBody{}
	if err := json.NewDecoder(r.Body).Decode(body); err != nil {
		errorResp := UpdateMember400Response{
			ProblemDetails: server.NewBadRequestResponse(err.Error(), r.URL.Path),
		}
		if err := errorResp.VisitUpdateMemberResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}
	input.Role = body.Role

	// Execute
	result, err := h.updateMember.Execute(ctx, input)
	if err != nil {
		errorResp := UpdateMember500Response{
			ProblemDetails: server.NewInternalServerErrorResponse(err.Error(), r.URL.Path),
		}
		if err := errorResp.VisitUpdateMemberResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}

	// Map output to response
	response := UpdateMember200Response{}
	response.Data = result.Data

	if err := response.VisitUpdateMemberResponse(w); err != nil {
		fmt.Fprintf(w, "error writing response: %v", err)
	}
}